requirements.txt
hashlib
figlet